Inference Pipeline §1.2 (Filters and Refraction) Another protective layer has been added, now concerning signal superimposition:
The signal_propagation_length type in all TOML configs and Rust structures is now u8 — we have eliminated the hole with “silent truncation” of huge u16 values when loading into the GPU.

A new check_single_spike_in_flight test has been added to Validator. Genesis Baker now strictly guarantees the invariant §1.6: refractory_period > signal_propagation_length. If the catfish's “sleep” period is shorter than the length of the impulse flying in the axon, there is a risk of two active “tails” on one wire at the same time. Baker will now reject such a config.

The documentation in spec 05 has been updated — errors with bit layout have been fixed (>> 16 changed to >> 10 according to constants)

+++ genesis-core/src/config/blueprints.rs
@@ -27,7 +27,7 @@ pub struct NeuronType {
     pub conduction_velocity: u16,
     
     #[serde(default = "default_propagation_length")]
-    pub signal_propagation_length: u16,
+    pub signal_propagation_length: u8,
 
     // --- Рост аксона (Steering) ---
     #[serde(default = "default_axon_growth_step")]
@@ -111,7 +111,7 @@ pub struct NeuronType {
 }
 
 // Дефолтные значения для опциональных полей
-fn default_propagation_length() -> u16 { 10 }
+fn default_propagation_length() -> u8 { 10 }
 fn default_axon_growth_step() -> u16 { 12 }
 fn default_steering_fov() -> f32 { 60.0 }
 fn default_steering_radius() -> f32 { 100.0 }
